+++ src/main/java/example/com/moneymergebe/domain/record/service/RecordService.java
@@ -154,7 +154,7 @@ public class RecordService {
         Category category = findCategory(req.getCategoryId());
 
         String imageUrl = record.getImage();
-        if (multipartFile != null && !multipartFile.isEmpty()) { // 새로 입력한 이미지 파일이 있는 경우
+        if (imageUrl != null && multipartFile != null && !multipartFile.isEmpty()) { // 새로 입력한 이미지 파일이 있는 경우
             if (s3Util.exists(imageUrl, FilePath.RECORD)) { // 기존 이미지가 존재하는 경우
                 s3Util.deleteFile(imageUrl, FilePath.RECORD); // 기존 이미지 삭제
             }
